Тип данных для скидки изменён с Double на BigDecimal для того, чтобы не проводить конвертацию в BigDecimal во время вычисления скидки.

+++ src/main/kotlin/ru/evreke/demo/entity/Category.kt
@@ -1,12 +1,13 @@
 package ru.evreke.demo.entity
 
+import java.math.BigDecimal
 import javax.persistence.*
 
 @Entity
 @Table(name = "categories")
 data class Category(
     var title: String?,
-    var discount: Double?
+    var discount: BigDecimal?
 ) {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
